conguguration all DB, no collector impl yet
