+++ src/main/scala/lambdaone/kratia/protocol/KratiaService.scala
@@ -26,7 +26,7 @@ import org.http4s.server.middleware._
 import cats.effect.implicits._
 import doobie.util.{Get, Put, Read}
 import doobie.util.transactor.Transactor
-import lambdaone.toolbox.sql.CrudPickSqlRegistry
+import lambdaone.toolbox.sql.{CrudPickSqlCollector, CrudPickSqlRegistry}
 
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration._
@@ -34,7 +34,7 @@ import scala.concurrent.duration._
 /**
   * Current unique community 19ce7b9b-a4da-4f9c-9838-c04fcb0ce9db
   */
-case class KratiaInMem(
+case class KratiaService(
                         uniqueGen: UniqueGen[IO, UUID],
                         registry: Registry[IO, UUID, MemberData],
                         collector: Collector[IO, UUID, BinaryProposal, String]
@@ -140,69 +140,4 @@ case class KratiaInMem(
         } yield ok
       }
   }
-}
-
-object KratiaInMem {
-
-  def inMem: IO[KratiaInMem] =
-    for {
-      registry <- buildInMemRegistry
-      collector <- buildInMemCollector
-    } yield KratiaInMem(UniqueGen.UniqueGenUUID, registry, collector)
-
-  def buildInMemRegistry: IO[Registry[IO, UUID, MemberData]] =
-    for {
-      store <- Ref.of[IO, Map[(UUID, UUID), MemberData]](Map.empty)
-      crud = CRUDPickInMem(store)
-      reg = RegistryCRUD(crud)
-    } yield reg
-
-  def buildInMemCollector: IO[Collector[IO, UUID, BinaryProposal, String]] =
-    for {
-      store <- Ref.of[IO, Map[UUID, BoxData[UUID, BinaryProposal, String]]](Map.empty)
-      clock = Clock.create[IO]
-      crud = CRUDPickInMem(store)
-      collector = CollectorCRUD[IO, UUID, BinaryProposal, String](clock, crud, UniqueGen.UniqueGenUUID)
-    } yield collector
-}
-
-object KratiaInDb {
-
-  implicit val cs = IO.contextShift(ExecutionContext.global)
-
-  implicit val xa = Transactor
-    .fromDriverManager[IO]("org.h2.Driver", "jdbc:h2:mem:kratia;DB_CLOSE_DELAY=-1")
-
-  implicit val uuidGet: Get[UUID] = Get[String].map(UUID.fromString(_))
-  implicit val uuidPut: Put[UUID] = Put[String].contramap(_.toString)
-
-  def buildDbRegistry: IO[Registry[IO, UUID, MemberData]] =
-    for {
-      _ <- CrudPickSqlRegistry.createTable[IO]
-      query: CRUDPick[IO, (UUID, UUID), MemberData] = CrudPickSqlRegistry[IO, UUID, MemberData]
-      reg: RegistryCRUD[IO, UUID, MemberData] = RegistryCRUD(query)
-    } yield reg
-
-  def buildInMemRegistry: IO[Registry[IO, UUID, MemberData]] =
-    for {
-      store <- Ref.of[IO, Map[(UUID, UUID), MemberData]](Map.empty)
-      crud = CRUDPickInMem(store)
-      reg = RegistryCRUD(crud)
-    } yield reg
-
-  def inDb: IO[KratiaInMem] =
-    for {
-      registry <- buildDbRegistry
-      collector <- buildInMemCollector
-    } yield KratiaInMem(UniqueGen.UniqueGenUUID, registry, collector)
-
-  // This is still in memory
-
-  def buildInMemCollector: IO[Collector[IO, UUID, BinaryProposal, String]] =
-    for {
-      store <- Ref.of[IO, Map[UUID, BoxData[UUID, BinaryProposal, String]]](Map.empty)
-      clock = Clock.create[IO]
-      crud = CRUDPickInMem(store)
-      collector = CollectorCRUD[IO, UUID, BinaryProposal, String](clock, crud, UniqueGen.UniqueGenUUID)
-    } yield collector
 }
